+++ src/geefetch/data/get.py
@@ -34,6 +34,7 @@ from ..utils.progress_multiprocessing import (
     QueuedProgress,
 )
 from ..utils.rasterio import create_vrt
+from ..utils.spectral_indices.spectral_index import SpectralIndex
 from .process import (
     geofile_is_clean,
     merge_tracked_geojson,
@@ -119,7 +120,10 @@ def download_chip(
     **kwargs: Any,
 ) -> Path:
     """Download a specific chip of data from the satellite."""
-    bands = selected_bands if selected_bands is not None else satellite.default_selected_bands
+    bands = list(selected_bands) if selected_bands is not None else satellite.default_selected_bands
+    spectra_indices: list[SpectralIndex] | None = data_get_kwargs.get("spectral_indices")
+    if spectra_indices:
+        bands += [index.name for index in spectra_indices]
     if out.exists():
         log.debug(f"Found feature chip [cyan]{out}[/]")
         if not geofile_is_clean(out):
@@ -536,6 +540,7 @@ def download_s1(
     terrain_normalization_config: TerrainNormalizationConfig | None = None,
     orbit: S1Orbit = S1Orbit.ASCENDING,
     resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+    spectral_indices: list[SpectralIndex] | None = None,
 ) -> None:
     """Download Sentinel-1 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `s1.vrt` is written to combine all the chips.
@@ -583,6 +588,8 @@ def download_s1(
         The resampling method to use when reprojecting images.
         Can be BILINEAR, BICUBIC or NEAREST.
         Defaults to ResamplingMethod.BILINEAR.
+    spectral_indices : list[SpectralIndex] | None
+        List of indices to calculate and add as bands of the downloaded images. Defaults to None
     """
 
     download_selected_bands: list[str] | None
@@ -618,6 +625,7 @@ def download_s1(
             "resolution": resolution,
             "speckle_filter_config": speckle_filter_config,
             "terrain_normalization_config": terrain_normalization_config,
+            "spectral_indices": spectral_indices,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
         as_time_series=(composite_method == CompositeMethod.TIMESERIES),
@@ -641,6 +649,7 @@ def download_s2(
     cloudless_portion: int = 60,
     cloud_prb_thresh: int = 40,
     resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+    spectral_indices: list[SpectralIndex] | None = None,
 ) -> None:
     """Download Sentinel-2 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `s2.vrt` is written to combine all the chips.
@@ -687,6 +696,8 @@ def download_s2(
         The resampling method to use when reprojecting images.
         Can be BILINEAR, BICUBIC or NEAREST.
         Defaults to ResamplingMethod.BILINEAR.
+    spectral_indices : list[SpectralIndex] | None
+        List of indices to calculate and add as bands of the downloaded images. Defaults to None
     """
     download(
         data_dir=data_dir,
@@ -708,6 +719,7 @@ def download_s2(
             "dtype": dtype,
             "resampling": resampling,
             "resolution": resolution,
+            "spectral_indices": spectral_indices,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
         as_time_series=(composite_method == CompositeMethod.TIMESERIES),
@@ -810,6 +822,7 @@ def download_landsat8(
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
     resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+    spectral_indices: list[SpectralIndex] | None = None,
 ) -> None:
     """Download Landsat 8 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `landsat8.vrt` is written to combine all the chips.
@@ -851,6 +864,8 @@ def download_landsat8(
         The resampling method to use when reprojecting images.
         Can be BILINEAR, BICUBIC or NEAREST.
         Defaults to ResamplingMethod.BILINEAR.
+    spectral_indices : list[SpectralIndex] | None
+        List of indices to calculate and add as bands of the downloaded images. Defaults to None
     """
     download(
         data_dir=data_dir,
@@ -870,6 +885,7 @@ def download_landsat8(
             "dtype": dtype,
             "resampling": resampling,
             "resolution": resolution,
+            "spectral_indices": spectral_indices,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
         as_time_series=(composite_method == CompositeMethod.TIMESERIES),
@@ -893,6 +909,7 @@ def download_palsar2(
     orbit: P2Orbit = P2Orbit.DESCENDING,
     resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
     refined_lee: bool = True,
+    spectral_indices: list[SpectralIndex] | None = None,
 ) -> None:
     """Download Palsar 2 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `palsar2.vrt` is written to combine all the chips.
@@ -939,6 +956,8 @@ def download_palsar2(
     refined_lee : bool
         Whether to apply the Refined Lee filter to reduce speckle noise.
         Defaults to True.
+    spectral_indices : list[SpectralIndex] | None
+        List of indices to calculate and add as bands of the downloaded images. Defaults to None
     """
     download(
         data_dir=data_dir,
@@ -960,6 +979,7 @@ def download_palsar2(
             "resampling": resampling,
             "resolution": resolution,
             "refined_lee": refined_lee,
+            "spectral_indices": spectral_indices,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
         as_time_series=(composite_method == CompositeMethod.TIMESERIES),